treewide: rename query_error.rs to execution_error.rs
QueryError was replaced with ExecutionError in Rust driver 1.0.
