+++ scylla-rust-wrapper/src/future.rs
@@ -2,8 +2,8 @@ use crate::argconv::*;
 use crate::cass_error::CassError;
 use crate::cass_error::CassErrorMessage;
 use crate::cass_error::ToCassError;
+use crate::execution_error::CassErrorResult;
 use crate::prepared::CassPrepared;
-use crate::query_error::CassErrorResult;
 use crate::query_result::CassResult;
 use crate::types::*;
 use crate::uuid::CassUuid;
