+++ scylla-rust-wrapper/src/lib.rs
@@ -17,6 +17,7 @@ pub mod cluster;
 pub mod collection;
 pub mod date_time;
 pub mod exec_profile;
+pub mod execution_error;
 mod external;
 pub mod future;
 pub mod inet;
@@ -26,7 +27,6 @@ mod logging;
 pub mod metadata;
 pub mod misc;
 pub mod prepared;
-pub mod query_error;
 pub mod query_result;
 pub mod retry_policy;
 pub mod session;
